Separate into two components
