--- packages/manager/src/features/ObjectStorage/ObjectUpload/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ObjectUpload';
