+++ packages/manager/src/features/ObjectStorage/ObjectUploader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ObjectUploader';
